refactor(dashboard): H1 — extract chat_completion + helper cluster → routes/chat.py (SR-59)
codec_dashboard.py: 2,022 → 1,310 LOC (-712 / -35.2%).
Cumulative since pre-B6 baseline: 3,912 → 1,310 (-66.5%).

The largest remaining endpoint (the /api/chat LLM handler) plus its tightly-
coupled private helpers move as one verbatim cluster into routes/chat.py:
  - _fetch_url_content        URL strip-and-fetch for inline context
  - _enrich_messages          memory / URL / web-search context injection
  - CHAT_SKILL_ALLOWLIST       skills a chat message may auto-fire (53 entries)
  - _try_skill                pre-LLM hijack (shared w/ /api/command — re-exported)
  - _try_skill_by_name        post-LLM [SKILL:] resolver + calculator fallback
  - _chat_vision_response     image → vision-model branch (A-11 pending)
  - _build_chat_system_prompt override + step-budget + observer suffixes
  - chat_completion           the POST /api/chat endpoint

Behavior preserved verbatim (script-extracted, not retyped). All safety
surfaces intact and pinned by a new regression test:
  * pre-LLM hijack consumes 1 step + gates destructive skills on
    codec_consent.chat_consent_ok
  * post-LLM [SKILL:] tags stay allowlist-gated + budget-gated + DROPPED
    (never leaked raw) on failure — both stream + non-stream paths
  * the <think>/[SKILL:] streaming token machine stays codec_chat_stream.SkillTagBuffer

Back-compat: codec_dashboard re-exports all 7 helpers identity-equal to
routes.chat (verified in test), so /api/command's _try_skill call and the
existing test surface (codec_dashboard.CHAT_SKILL_ALLOWLIST, _build_chat_system_prompt,
etc.) keep working untouched. CHAT_SYSTEM_PROMPT stays in codec_dashboard
(routes.prompts also imports it) and is lazy-imported in routes/chat.py to
avoid a load-time cycle.

What stays in codec_dashboard: /api/command (safety-critical), the auto-
escalate classifier cluster (Step 10, independent of chat_completion),
/api/services/status, page renderers, startup/shutdown hooks, _bg_* daemons.

Test surface:
  - new TestH1ChatExtraction (6 pins): endpoint registered, router export,
    7-helper identity-equal re-export, safety-surfaces-preserved scan,
    dashboard-doesn't-redefine, LOC floor < 1,400
  - test_dashboard_llm: codec_llm.call >=3→>=2 (chat non-stream moved),
    /chat/completions ==2→==1 (only _warmup_vision left), stream/KEEPALIVE
    scan retargeted codec_dashboard → routes/chat
  - test_a12_invariant: routes/chat.py allowlisted (chat vision branch, A-11 pending)
  - test_route_extractions_d: lazy-import pin retargeted to routes/chat.py
  - drive-by ruff F401: dropped now-unused `secrets` + SkillTagBuffer/SKILL_TAG_RE
    imports from codec_dashboard

Full suite: 2,052 passed / 77 skipped (was 2,046 in G). +6 net.
