+++ codec_dashboard.py
@@ -5,7 +5,6 @@ import time
 import hmac
 import threading
 import asyncio
-import secrets
 from datetime import datetime, timedelta
 
 from fastapi import FastAPI, Request
@@ -33,8 +32,8 @@ from routes._shared import (
 # Audit emits route through the unified log_event adapter (real, not no-op)
 # per docs/PHASE1-STEP1-DESIGN.md.
 from codec_audit import log_event  # STEP_BUDGET_EXHAUSTED moved with _StepBudget to codec_chat_pipeline (B6-P2)
-from codec_chat_stream import SkillTagBuffer, SKILL_TAG_RE  # A-6 (PR-3D-c)
-import codec_llm  # A-12 (PR-3E-dashboard)
+# H1 / SR-59: SkillTagBuffer / SKILL_TAG_RE moved with chat_completion to routes/chat.py.
+import codec_llm  # A-12 (PR-3E-dashboard) — still used by the command Flash + classifier
 
 # E2 / SR-48: pydantic.BaseModel + Field moved with HealthResponse to
 # routes/health.py — the model was the only consumer.
@@ -360,6 +359,19 @@ from routes.cdp import router as cdp_router
 # G-series (SR-57..58): cross-source memory search + Pilot proxy.
 from routes.memory_search import router as memory_search_router
 from routes.pilot_proxy import router as pilot_proxy_router
+# H1 / SR-59: chat handler (POST /api/chat) + its helper cluster. The helpers
+# are re-exported back here (below) for the command-handler caller + the
+# existing test surface (codec_dashboard.CHAT_SKILL_ALLOWLIST etc.) — identity-equal.
+from routes.chat import router as chat_router
+from routes.chat import (  # noqa: F401  (back-compat re-exports)
+    CHAT_SKILL_ALLOWLIST,
+    _build_chat_system_prompt,
+    _chat_vision_response,
+    _enrich_messages,
+    _fetch_url_content,
+    _try_skill,
+    _try_skill_by_name,
+)
 # Phase 2 Step 6 — Trigger System PWA endpoints (auth-gated by /api/* middleware).
 try:
     from routes.triggers import router as triggers_router
@@ -396,6 +408,7 @@ app.include_router(web_search_router)
 app.include_router(cdp_router)
 app.include_router(memory_search_router)
 app.include_router(pilot_proxy_router)
+app.include_router(chat_router)
 if _has_triggers:
     app.include_router(triggers_router)
 
@@ -779,259 +792,11 @@ async def vibe_page():
 # (Forge endpoint moved to routes/skills.py)
 
 
-def _fetch_url_content(url: str, max_chars: int = 8000) -> str:
-    """Fetch a URL and return stripped text content."""
-    try:
-        import httpx
-        from html.parser import HTMLParser
-
-        class _Stripper(HTMLParser):
-            def __init__(self):
-                super().__init__()
-                self._skip = False
-                self.chunks = []
-            def handle_starttag(self, tag, attrs):
-                if tag in ('script', 'style', 'nav', 'footer'):
-                    self._skip = True
-            def handle_endtag(self, tag):
-                if tag in ('script', 'style', 'nav', 'footer'):
-                    self._skip = False
-            def handle_data(self, data):
-                if not self._skip:
-                    stripped = data.strip()
-                    if stripped:
-                        self.chunks.append(stripped)
-
-        r = httpx.get(url, timeout=15, follow_redirects=True,
-                       headers={"User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/131.0.0.0 Safari/537.36",
-                                "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8"})
-        if 'text/html' in r.headers.get('content-type', ''):
-            parser = _Stripper()
-            parser.feed(r.text)
-            text = ' '.join(parser.chunks)
-        else:
-            text = r.text
-        return text[:max_chars]
-    except Exception as e:
-        log.warning(f"URL fetch failed ({url}): {e}")
-        return ""
-
-
-def _enrich_messages(messages: list, config: dict, force_search: bool = False) -> list:
-    """
-    Auto-detect URLs, search intent, and memory recall in the last user message.
-    Injects context messages before the last user message when content is found.
-    force_search=True bypasses intent detection and always searches.
-    Returns a (possibly modified) copy of the messages list.
-    """
-    import re as _re
-    if not messages:
-        return messages
-
-    # Find last user message
-    last_user_idx = None
-    for i in range(len(messages) - 1, -1, -1):
-        if messages[i].get("role") == "user":
-            last_user_idx = i
-            break
-    if last_user_idx is None:
-        return messages
-
-    last_text = messages[last_user_idx].get("content", "")
-    if not isinstance(last_text, str):
-        return messages
-
-    context_parts = []
-    memory_parts = []
-
-    # ── Memory recall ──────────────────────────────────────────────────────────
-    # Inject relevant memory context from ALL sources (voice, chat, vibe) for
-    # full cross-session recall.
-    lower = last_text.lower()
-    memory_triggers = [
-        'remember', 'recall', 'earlier', 'before', 'last time',
-        'previously', 'we talked', 'we discussed', 'you said',
-        'did i', 'did we', 'have i', 'have we', 'my previous',
-        'past conversation', 'history', 'do you know my',
-        'what was', 'what did', 'when did',
-    ]
-    has_memory_trigger = any(t in lower for t in memory_triggers)
-
-    # 1. Voice memory (FTS5 via CodecMemory) — always inject recent, targeted on trigger
-    try:
-        from codec_memory import CodecMemory
-        mem = CodecMemory()
-        if has_memory_trigger:
-            mem_context = mem.get_context(last_text, n=8)
-            if mem_context:
-                memory_parts.append(f"[MEMORY — RELEVANT PAST CONVERSATIONS (VOICE)]\n{mem_context}\n[END MEMORY]")
-                log.info(f"Memory recall injected (voice targeted): {len(mem_context)} chars")
-        recent = mem.search_recent(days=3, limit=5)
-        if recent:
-            lines = ["[RECENT MEMORY — VOICE (LAST 3 DAYS)]"]
-            for r in recent:
-                ts = r["timestamp"][:16].replace("T", " ")
-                snippet = r["content"][:200].replace("\n", " ")
-                lines.append(f"  [{ts}] {r['role'].upper()}: {snippet}")
-            lines.append("[END RECENT MEMORY]")
-            memory_parts.append("\n".join(lines))
-            log.info(f"Recent memory injected: {len(recent)} messages")
-    except Exception as e:
-        log.warning(f"Memory enrichment (voice) failed: {e}")
-
-    # 2. Dashboard chat history (qchat.db) — targeted search on trigger, recent always
-    try:
-        from routes.qchat import qchat_db as _qchat_db; _qc = _qchat_db()
-        if has_memory_trigger:
-            keyword = f"%{last_text[:80]}%"
-            qrows = _qc.execute(
-                "SELECT role, content, timestamp FROM qchat_messages "
-                "WHERE content LIKE ? COLLATE NOCASE ORDER BY id DESC LIMIT 6",
-                (keyword,)
-            ).fetchall()
-            if qrows:
-                lines = ["[MEMORY — RELEVANT PAST CHATS]"]
-                for r in qrows:
-                    ts = (r[2] or "")[:16].replace("T", " ")
-                    snippet = (r[1] or "")[:200].replace("\n", " ")
-                    lines.append(f"  [{ts}] {(r[0] or '').upper()}: {snippet}")
-                lines.append("[END MEMORY]")
-                memory_parts.append("\n".join(lines))
-                log.info(f"Memory recall injected (chat targeted): {len(qrows)} msgs")
-        # Recent chat messages for continuity
-        qrecent = _qc.execute(
-            "SELECT role, content, timestamp FROM qchat_messages ORDER BY id DESC LIMIT 5"
-        ).fetchall()
-        if qrecent:
-            lines = ["[RECENT MEMORY — CHAT]"]
-            for r in qrecent:
-                ts = (r[2] or "")[:16].replace("T", " ")
-                snippet = (r[1] or "")[:200].replace("\n", " ")
-                lines.append(f"  [{ts}] {(r[0] or '').upper()}: {snippet}")
-            lines.append("[END RECENT MEMORY]")
-            memory_parts.append("\n".join(lines))
-            log.info(f"Recent chat memory injected: {len(qrecent)} messages")
-    except Exception as e:
-        log.warning(f"Memory enrichment (chat) failed: {e}")
-
-    # 3. Vibe IDE history (vibe.db) — targeted search on trigger only (less relevant day-to-day)
-    if has_memory_trigger:
-        try:
-            from routes.vibe import vibe_db as _vibe_db; _vc = _vibe_db()
-            keyword = f"%{last_text[:80]}%"
-            vrows = _vc.execute(
-                "SELECT role, content, timestamp FROM vibe_messages "
-                "WHERE content LIKE ? COLLATE NOCASE ORDER BY id DESC LIMIT 4",
-                (keyword,)
-            ).fetchall()
-            if vrows:
-                lines = ["[MEMORY — RELEVANT VIBE/CODE CONVERSATIONS]"]
-                for r in vrows:
-                    ts = (r[2] or "")[:16].replace("T", " ")
-                    snippet = (r[1] or "")[:200].replace("\n", " ")
-                    lines.append(f"  [{ts}] {(r[0] or '').upper()}: {snippet}")
-                lines.append("[END MEMORY]")
-                memory_parts.append("\n".join(lines))
-                log.info(f"Memory recall injected (vibe targeted): {len(vrows)} msgs")
-        except Exception as e:
-            log.warning(f"Memory enrichment (vibe) failed: {e}")
-
-    # ── URL detection ──────────────────────────────────────────────────────────
-    urls = _re.findall(r'https?://[^\s\)\]>,"\']+', last_text)
-    for url in urls[:3]:  # cap at 3 URLs per message
-        content = _fetch_url_content(url)
-        if content:
-            context_parts.append(f"[URL CONTENT: {url}]\n{content}\n[END URL CONTENT]")
-            log.info(f"Chat URL fetched: {url} ({len(content)} chars)")
-
-    # ── Search intent detection ────────────────────────────────────────────────
-    search_triggers = [
-        'search for', 'search the web', 'google', 'look up', 'find out',
-        'what is the latest', 'current news', 'recent', 'today\'s', 'right now',
-        'who won', 'stock price', 'weather in', 'news about'
-    ]
-    lower = last_text.lower()
-    should_search = (any(t in lower for t in search_triggers) or force_search) and not urls
-    if should_search:
-        try:
-            import sys
-            import os as _os
-            repo_dir = _os.path.dirname(_os.path.abspath(__file__))
-            if repo_dir not in sys.path:
-                sys.path.insert(0, repo_dir)
-            from codec_search import search, format_results
-            results = search(last_text, max_results=5)
-            if results:
-                context_parts.append(f"[WEB SEARCH RESULTS]\n{format_results(results, max_snippets=5)}\n[END WEB SEARCH RESULTS]")
-                log.info(f"Chat search injected for: {last_text[:80]}")
-        except Exception as e:
-            log.warning(f"Chat search failed: {e}")
-
-    if not context_parts and not memory_parts:
-        return messages
-
-    enriched = list(messages)
-
-    # Inject memory + other context as a single user message (hidden context)
-    # Using role "user" with clear [INTERNAL] framing so local LLMs don't choke on mid-conversation "system" role
-    all_context = memory_parts + context_parts
-    if all_context:
-        prefix = ("(INTERNAL CONTEXT — do not echo this block. Use it to inform your answer naturally. "
-                   "Never show raw [MEMORY] or [RECENT MEMORY] tags to the user.)\n\n")
-        context_msg = {"role": "user", "content": prefix + "\n\n".join(all_context)}
-        enriched.insert(last_user_idx, context_msg)
-
-    return enriched
-
-
+# H1 / SR-59: def _fetch_url_content → moved to routes/chat.py
+# H1 / SR-59: def _enrich_messages → moved to routes/chat.py
 # F6 web_search → moved to routes/*.py
 # ── Chat Tool Calling: safe skills available from Chat ──
-CHAT_SKILL_ALLOWLIST = {
-    # Core utilities
-    "calculator", "weather", "web_search", "bitcoin_price",
-    "system", "network_info", "memory_search", "time",
-    "timer", "translate", "file_search", "notes",
-    "reminders", "clipboard", "password_generator",
-    "qr_generator", "json_formatter", "pomodoro",
-    # Terminal / shell (goes through is_dangerous safety check)
-    "terminal",
-    # File operations (read, write, append, list — path-restricted)
-    "file_ops",
-    # NOTE: python_exec is intentionally NOT on this allowlist (audit C3).
-    # It stays a local skill but is no longer auto-firable from a chat message
-    # (pre-LLM hijack / post-LLM [SKILL:...] tag both gate on this set), so an
-    # injection-style chat message can't drive arbitrary code execution.
-    # SKILL_MCP_EXPOSE=False already keeps it off MCP.
-    # Google services
-    "google_calendar", "google_gmail", "google_docs",
-    "google_drive", "google_sheets", "google_keep",
-    "google_tasks", "google_slides",
-    # Browser control
-    "chrome_automate", "chrome_click_cdp", "chrome_read",
-    "chrome_extract", "chrome_fill", "chrome_scroll",
-    "chrome_open", "chrome_close", "chrome_tabs", "chrome_search",
-    # System control (volume, brightness, apps — NO mouse_control)
-    "screenshot_text", "app_switch",
-    "brightness", "volume_brightness", "process_manager",
-    "ax_control",
-    # PM2 service management
-    "pm2_control",
-    # Smart home & media
-    "philips_hue", "music",
-    # Self-improvement & meta
-    "ai_news_digest", "scheduler",
-    # Skill creation & delegation
-    # re-audit (CHAIN-002): skill_forge writes forged code to disk WITHOUT the
-    # review gate, so it must not be auto-firable from a chat [SKILL:...] tag —
-    # skill creation goes through create_skill's review-and-approve flow only
-    # (PR-1B). ask_codec_to_build had no backing skill file (stale entry).
-    "create_skill", "delegate",
-    # Phase 2 Step 7 — end-of-day shift report (read-only, no destructive side effects)
-    "shift_report",
-    # Phase 2 Step 6 — first declarative trigger (clipboard URL → web_fetch).
-    # Read-only network fetch, gated by codec_ask_user.ask consent on auto-fire.
-    "clipboard_url_fetch",
-}
+# H1 / SR-59: CHAT_SKILL_ALLOWLIST = { → moved to routes/chat.py
 
 # ---------------------------------------------------------------------------
 # Chat System Prompt — gives the LLM identity, context, and skill awareness
@@ -1099,104 +864,8 @@ from codec_chat_pipeline import (  # noqa: E402,F401  (back-compat re-exports)
 )
 
 
-def _try_skill(user_text: str):
-    """Check if user_text matches a skill. Returns (skill_name, result) or (None, None).
-    Skips skill matching for conversational messages to prevent false triggers."""
-    if _is_conversational(user_text):
-        return None, None
-    try:
-        from codec_dispatch import check_skill, run_skill
-        skill = check_skill(user_text)
-        if skill and skill.get("name") in CHAT_SKILL_ALLOWLIST:
-            # re-audit A2: destructive skills need explicit consent (reuses the
-            # AskUserQuestion PWA panel; blocks this worker thread until answered).
-            import codec_consent
-            if not codec_consent.chat_consent_ok(skill["name"], user_text):
-                return skill["name"], (
-                    f"⚠ '{skill['name']}' is a destructive operation and wasn't "
-                    "confirmed — skipped."
-                )
-            result = run_skill(skill, user_text, app="CODEC Chat")
-            if result is not None:
-                return skill["name"], str(result)
-    except Exception as e:
-        log.warning(f"[Chat] Skill check error: {e}")
-    return None, None
-
-
-def _try_skill_by_name(name: str, query: str):
-    """Execute a specific skill by name (for LLM-routed skill calls).
-
-    For calculator specifically: LLMs often pass natural-language descriptions
-    like "sum of Facebook (4900 + 6100), LinkedIn ...". The calculator skill
-    can't parse that. We try the raw query first, then fall back to extracting
-    every number out of the string and summing/computing locally so the user
-    always gets a number instead of a raw [SKILL:...] tag leaking through.
-    """
-    try:
-        from codec_dispatch import run_skill
-        skill = {"name": name}
-        # re-audit A2: a destructive skill emitted via a post-LLM [SKILL:...] tag
-        # (the prompt-injection vector) needs explicit consent before it runs —
-        # reuses the AskUserQuestion PWA panel. Blocks until answered.
-        import codec_consent
-        if not codec_consent.chat_consent_ok(name, query):
-            return name, (
-                f"⚠ '{name}' is a destructive operation and wasn't confirmed — skipped."
-            )
-        result = run_skill(skill, query, app="CODEC Chat (LLM-routed)")
-        if result is not None:
-            return name, str(result)
-    except Exception as e:
-        log.warning(f"[Chat] LLM skill route error ({name}): {e}")
-
-    # Calculator-specific fallback: rescue messy LLM-routed inputs like
-    #   "sum of Facebook (4900 + 6100), LinkedIn (4127 + 3900), ..."
-    # The LLM almost always means "give me the total" → extract every number
-    # and sum. If the query plainly contains a single arithmetic expression,
-    # we eval that instead.
-    if name == "calculator":
-        try:
-            import re as _re_calc
-            q_lower = query.lower()
-            # Detect a clean arithmetic expression like "47*89" with no other words
-            stripped = _re_calc.sub(r"[^0-9+\-*/().\s]", "", query).strip()
-            stripped = _re_calc.sub(r"\s+", "", stripped)
-            looks_like_clean_expression = (
-                stripped
-                and _re_calc.fullmatch(r"[0-9+\-*/().]+", stripped)
-                and _re_calc.search(r"[+\-*/]", stripped)
-            )
-            if looks_like_clean_expression:
-                try:
-                    val = eval(stripped, {"__builtins__": {}}, {})  # noqa: S307
-                    return name, f"{val:,}"
-                except Exception:
-                    pass
-
-            # Otherwise: pull every number out and decide an op based on intent
-            nums = [float(n) for n in _re_calc.findall(r"\d+(?:\.\d+)?", query)]
-            if len(nums) >= 2:
-                # Default to sum (covers grand total / how many / count / etc).
-                # If user said "product" / "multiply" / "times" → multiply.
-                if any(_re_calc.search(rf"\b{kw}\b", q_lower)
-                       for kw in ("product", "multiply", "multiplied", "times")):
-                    val = 1.0
-                    for n in nums:
-                        val *= n
-                else:
-                    val = sum(nums)
-                # Format integer-clean if no decimals were involved
-                val_int = int(val)
-                if val == val_int:
-                    return name, f"{val_int:,}"
-                return name, f"{val:,.2f}"
-        except Exception as e:
-            log.warning(f"[Chat] calculator fallback failed: {e}")
-
-    return name, None
-
-
+# H1 / SR-59: def _try_skill → moved to routes/chat.py
+# H1 / SR-59: def _try_skill_by_name → moved to routes/chat.py
 # ── Phase 3 Step 10 — Auto-escalation classifier ──────────────────────────
 
 _AUTO_ESCALATE_SYSTEM_PROMPT = """You are CODEC's chat-input classifier. \
@@ -1322,390 +991,9 @@ def _should_escalate_to_project(user_text: str, session_id: str) -> dict:
     }
 
 
-def _chat_vision_response(body: dict, messages: list):
-    """If the request carries images, route to the vision model and return the
-    response dict; else return None. Fix #8 (intra-file CC reduction):
-    extracted verbatim from chat_completion, behavior-preserving. The inline
-    vision POST is an A-11-pending site and stays in codec_dashboard."""
-    images = body.get("images", [])
-    if not images:
-        return None
-    import requests as rq2
-    config2 = {}
-    try:
-        with open(CONFIG_PATH) as f:
-            config2 = json.load(f)
-    except (OSError, json.JSONDecodeError) as e:
-        log.warning(f"Config read failed; proceeding without overrides: {e}")
-    vision_url = config2.get("vision_base_url", "http://localhost:8083/v1")
-    vision_model = config2.get("vision_model", "mlx-community/Qwen2.5-VL-7B-Instruct-4bit")
-    # Build multimodal message: last user text + all images
-    last_text = ""
-    for m in reversed(messages):
-        if m.get("role") == "user" and isinstance(m.get("content"), str):
-            last_text = m["content"]
-            break
-    if not last_text:
-        last_text = "Describe and analyze this image in detail."
-    mm_content = []
-    for img_b64 in images:
-        mm_content.append({"type": "image_url", "image_url": {"url": f"data:image/jpeg;base64,{img_b64}"}})
-    mm_content.append({"type": "text", "text": last_text})
-    v_payload = {
-        "model": vision_model,
-        "messages": [{"role": "user", "content": mm_content}],
-        "max_tokens": 4000,
-        "temperature": 0.7
-    }
-    # re-audit N7: guard the vision-backend call + parse. This helper runs
-    # OUTSIDE chat_completion's try/except, so a non-200 / malformed response
-    # (model not loaded, OOM, timeout) previously surfaced as a raw 500 with no
-    # JSON body. Return a graceful 502 instead.
-    try:
-        vr = rq2.post(f"{vision_url}/chat/completions", json=v_payload,
-                      headers={"Content-Type": "application/json"}, timeout=120)
-        vr.raise_for_status()
-        vdata = vr.json()
-        vanswer = vdata["choices"][0]["message"]["content"].strip()
-    except Exception as e:
-        log.warning(f"[chat] vision backend call failed: {e}")
-        return JSONResponse(
-            {"error": f"Vision model unavailable: {type(e).__name__}"},
-            status_code=502,
-        )
-    import re as re2
-    vanswer = re2.sub(r'<think>[\s\S]*?</think>', '', vanswer).strip()
-    return {"response": vanswer, "model": vision_model}
-
-
-def _build_chat_system_prompt(config: dict, budget, has_attachment: bool,
-                              last_user_text: str) -> str:
-    """Build the chat system prompt: override-aware base + per-turn step-budget
-    warnings + attachment / content-rewrite / observer-injection suffixes.
-
-    Fix #8 (intra-file CC reduction): extracted verbatim from chat_completion;
-    behavior-preserving. `budget` is mutated exactly as before — warn_now() and
-    consume('llm_call') happen here, once, where they ran inline.
-    """
-    from datetime import datetime as _dt
-    # D4 / SR-45: helper moved to routes.prompts; lazy-import at call time.
-    from routes.prompts import _load_prompt_overrides
-    _overrides = _load_prompt_overrides()
-    _chat_prompt = _overrides.get("chat", CHAT_SYSTEM_PROMPT)
-    sys_prompt = _chat_prompt.format(date=_dt.now().strftime("%A, %B %d, %Y"))
-    if budget.warn_now():
-        sys_prompt += (
-            "\n\n⚠ 1 step remaining in this turn. Wrap up — do NOT "
-            "emit additional [SKILL:...] tags."
-        )
-    budget.consume("llm_call")
-    if budget.at_limit():
-        sys_prompt += (
-            "\n\n## Step Budget Exhausted\n"
-            "You've hit the per-turn step budget. Summarize what you "
-            "accomplished and any blockers in one short paragraph. "
-            "DO NOT emit [SKILL:...] tags or call additional tools."
-        )
-    if has_attachment:
-        sys_prompt += (
-            "\n\n## This Turn\n"
-            "The user has attached a file or image and its content is already "
-            "embedded in their message between [IMAGE ANALYSIS]/[DOCUMENT] markers. "
-            "Respond conversationally about the attached content. "
-            "DO NOT emit [SKILL:...] tool-calling tags in this response."
-        )
-    _u_text_lower = (last_user_text or "").lower()
-    _content_rewrite_intent = any(
-        kw in _u_text_lower for kw in (
-            "format my email", "format this email", "format my message",
-            "reformat", "rewrite", "reword", "redraft", "polish",
-            "proofread", "edit my email", "fix my email", "fix the grammar",
-            "make this sound", "translate this", "translate the following",
-            "draft a reply", "draft an email", "draft this",
-        )
-    )
-    if _content_rewrite_intent:
-        sys_prompt += (
-            "\n\n## This Turn\n"
-            "The user is asking you to generate or rewrite text directly "
-            "(format/edit/draft/translate/polish their email or message). "
-            "Respond with the rewritten content as plain prose. "
-            "DO NOT emit [SKILL:...] tool-calling tags in this response — "
-            "the answer IS the rewritten text, no tools needed."
-        )
-    try:
-        from codec_observer import maybe_inject_observation_summary
-        _obs_transport = "local" if "localhost" in (config.get("llm_base_url") or "") else "chat"
-        _obs_summary, _obs_reason = maybe_inject_observation_summary(
-            user_prompt=last_user_text or "",
-            transport=_obs_transport,
-            skill_name=None,           # post-LLM tag path, no skill resolved yet
-            skill_module=None,
-        )
-        if _obs_summary:
-            sys_prompt += f"\n\n{_obs_summary}"
-    except Exception as _e:
-        log.debug(f"[observer] injection failed (non-fatal): {_e}")
-    return sys_prompt
-
-
-@app.post("/api/chat")
-async def chat_completion(request: Request):
-    """Direct LLM chat with full context window + tool calling"""
-    from codec_metrics import metrics
-    metrics.inc("codec_chat_requests_total")
-    body = await request.json()
-    messages = body.get("messages", [])
-    if not messages:
-        return JSONResponse({"error": "No messages"}, status_code=400)
-
-    # Phase 1 Step 3 §3 — per-turn step budget. One counter for the
-    # entire request; consumed by skill_hijack, llm_call, and each
-    # post-LLM [SKILL:] tag resolution. Budget enforcement is non-
-    # blocking (each path still runs) but audit-event-emitting +
-    # warn-at-N-1 prompt suffix injection. See _StepBudget docstring.
-    _budget = _StepBudget(
-        route="chat",
-        correlation_id=secrets.token_hex(6),
-    )
-
-    # ── Tool Calling: check if last user message matches a skill ──
-    use_tools = body.get("tools", True)  # frontend can disable with tools:false
-    if use_tools:
-        last_user_text = ""
-        for m in reversed(messages):
-            if m.get("role") == "user" and isinstance(m.get("content"), str):
-                last_user_text = m["content"]
-                break
-        # ── Slash commands (BEFORE skill check / attachment check) ──
-        # Type /help, /skills, /cost, /version, /status, /who, /clear in chat
-        # to invoke meta-controls without an LLM round-trip. Slash dispatch
-        # runs first so /version still works even if the user has an image
-        # attached in the same turn.
-        if last_user_text:
-            try:
-                from codec_slash_commands import parse_slash, dispatch as slash_dispatch
-                parsed = parse_slash(last_user_text)
-            except Exception as e:
-                log.warning(f"slash parser unavailable: {e}")
-                parsed = None
-            if parsed is not None:
-                cmd_name, cmd_args = parsed
-                slash_md = await asyncio.to_thread(slash_dispatch, cmd_name, cmd_args)
-                log.info(f"[Chat] Slash /{cmd_name} handled ({len(slash_md)} chars)")
-                stream_mode = body.get("stream", False)
-                if stream_mode:
-                    from starlette.responses import StreamingResponse as _SlashSR
-                    async def _slash_stream():
-                        yield f"data: {json.dumps({'slash': cmd_name})}\n\n"
-                        yield f"data: {json.dumps({'token': slash_md})}\n\n"
-                        yield "data: [DONE]\n\n"
-                    return _SlashSR(_slash_stream(), media_type="text/event-stream")
-                return {"response": slash_md, "slash": cmd_name}
-
-        # Skip skill routing when the user attached a file / image — otherwise the
-        # IMAGE ANALYSIS / DOCUMENT context text triggers false-positive skill hits
-        # (e.g. a screenshot describing "system dashboard" routes to system_info).
-        # Bugfix 2026-04-16: image attachments were being hijacked by skill router.
-        has_attachment = last_user_text and (
-            "[IMAGE ANALYSIS" in last_user_text
-            or "[DOCUMENT:" in last_user_text
-            or "[END IMAGE]" in last_user_text
-            or "[END DOCUMENT]" in last_user_text
-        )
-        if last_user_text and not has_attachment:
-            skill_name, skill_result = await asyncio.to_thread(_try_skill, last_user_text)
-            if skill_result:
-                _budget.consume("skill_hijack")   # pre-LLM hijack consumes 1
-                log.info(f"[Chat] Skill '{skill_name}' handled: {skill_result[:80]}")
-                stream_mode = body.get("stream", False)
-                if stream_mode:
-                    from starlette.responses import StreamingResponse as _SkillSR
-                    # Return skill result as SSE stream (same format as LLM stream)
-                    async def _skill_stream():
-                        # Send skill indicator
-                        yield f"data: {json.dumps({'skill': skill_name})}\n\n"
-                        # Send the result as a single token, then LLM follow-up
-                        skill_prefix = f"**⚡ {skill_name}**: {skill_result}\n\n"
-                        yield f"data: {json.dumps({'token': skill_prefix})}\n\n"
-                        yield "data: [DONE]\n\n"
-                    return _SkillSR(_skill_stream(), media_type="text/event-stream")
-                else:
-                    return {"response": f"**⚡ {skill_name}**: {skill_result}", "skill": skill_name}
-
-    # Check for images — route to vision model (extracted, Fix #8)
-    vision_resp = _chat_vision_response(body, messages)
-    if vision_resp is not None:
-        return vision_resp
-
-    try:
-        config = {}
-        try:
-            with open(CONFIG_PATH) as f: config = json.load(f)
-        except (OSError, json.JSONDecodeError) as e:
-            log.warning(f"Config read failed; proceeding without overrides: {e}")
-        base_url = config.get("llm_base_url", "http://localhost:8083/v1")
-        model = config.get("llm_model", "mlx-community/Qwen3.6-35B-A3B-4bit")
-        # PR-2B (D-15 partial): keychain-aware live read.
-        from codec_config import get_llm_api_key as _kc_get_llm
-        api_key = _kc_get_llm()
-        kwargs = config.get("llm_kwargs", {})
-        # (A-12 PR-3E-chat-stream: the `import requests as rq` + `headers` here are
-        # gone — both chat POSTs now go through codec_llm, which builds its own.)
-        force_search = body.get("force_search", False)
-        messages = _enrich_messages(messages, config, force_search=bool(force_search))
-
-        # Build the system prompt (override + step-budget + attachment /
-        # content-rewrite / observer suffixes) — extracted to a helper for
-        # readability (Fix #8). Consumes the llm_call step budget internally.
-        sys_prompt = _build_chat_system_prompt(
-            config, _budget, has_attachment, last_user_text
-        )
-
-        # Prepend system message (or replace existing one)
-        if messages and messages[0].get("role") == "system":
-            messages[0]["content"] = sys_prompt + "\n\n" + messages[0]["content"]
-        else:
-            messages.insert(0, {"role": "system", "content": sys_prompt})
-
-        stream_mode = body.get("stream", False)
-        # Dashboard chat & Vibe benefit from thinking mode (deeper answers).
-        # Frontend can send thinking=false to override for speed.
-        thinking = body.get("thinking", True)
-
-        # A-12 (PR-3E-chat-stream): build the shared codec_llm args ONCE so the
-        # stream + non-stream branches can't drift. top_p/frequency_penalty are
-        # explicit but kwargs may override them (matches the old payload.update);
-        # enable_thinking is the codec_llm param applied last → frontend toggle
-        # wins (matches the old chat_template_kwargs assignment after the update).
-        _extra = {"top_p": 0.9, "frequency_penalty": 1.1,
-                  **{k: v for k, v in kwargs.items() if k != "chat_template_kwargs"}}
-        _common = dict(base_url=base_url, model=model, api_key=api_key,
-                       max_tokens=28000, temperature=0.7, enable_thinking=thinking,
-                       extra_kwargs=_extra, timeout=300)
-
-        if stream_mode:
-            # SSE streaming — keeps Cloudflare tunnel alive, sends tokens as they arrive
-            def _stream_gen():
-                # A-6 (PR-3D-c): the <think> + [SKILL:...] token machine lives in
-                # codec_chat_stream.SkillTagBuffer; _resolve_skill_tag (below) is
-                # injected (it runs the skill: budget + allowlist + dispatch).
-                # A-12 (PR-3E-chat-stream): the SSE POST + keepalive are now
-                # codec_llm.stream(keepalive=True); this generator just wires the
-                # raw tokens through the buffer and frames them.
-
-                def _frame(tok):
-                    return f"data: {json.dumps({'token': tok})}\n\n"
-
-                def _resolve_skill_tag(raw_tag):
-                    """Run the skill inline and return its string result.
-
-                    On any failure we DROP the tag (return empty string) instead
-                    of leaking the raw [SKILL:...] tag into the UI. The LLM's
-                    own follow-up prose usually contains the answer anyway.
-                    Bugfix 2026-04-26: previously returned raw_tag on failure,
-                    causing "[SKILL:calculator:sum of...]" to appear in chat.
-
-                    Phase 1 Step 3 §3 — each resolved tag consumes one step
-                    from the chat-turn budget. If exhausted, the tag is
-                    dropped (so the LLM doesn't continue burning steps);
-                    step_budget_exhausted audit was already emitted by
-                    _budget.consume.
-                    """
-                    m = SKILL_TAG_RE.search(raw_tag)
-                    if not m:
-                        return raw_tag  # not a skill tag at all — emit as-is
-                    if not _budget.consume("post_llm_skill_tag"):
-                        log.info("[Chat] step_budget exhausted — dropping [SKILL:...] tag")
-                        return raw_tag.replace(m.group(0), "")
-                    s_name, s_query = m.group(1), m.group(2)
-                    if s_name not in CHAT_SKILL_ALLOWLIST:
-                        log.info(f"[Chat] LLM tried disallowed skill {s_name!r} — dropping tag")
-                        return raw_tag.replace(m.group(0), "")
-                    try:
-                        _, s_result = _try_skill_by_name(s_name, s_query)
-                        if s_result:
-                            return raw_tag.replace(m.group(0), f"**{s_result}**")
-                        log.info(f"[Chat] Skill {s_name!r} returned None for {s_query[:60]!r} — dropping tag")
-                    except Exception as e:
-                        log.warning(f"[Chat] Skill {s_name!r} crashed: {e}")
-                    # Drop the tag silently — never leak raw [SKILL:...] to UI
-                    return raw_tag.replace(m.group(0), "")
-                buf = SkillTagBuffer(_resolve_skill_tag)
-                try:
-                    # codec_llm.stream yields raw content deltas (it owns the SSE
-                    # POST + data:/[DONE] parsing) and the KEEPALIVE sentinel on
-                    # empty thinking-chunks (keepalive=True) to hold the tunnel.
-                    for item in codec_llm.stream(messages, **_common, keepalive=True):
-                        if item is codec_llm.KEEPALIVE:
-                            yield ": keepalive\n\n"
-                            continue
-                        for s in buf.feed(item):
-                            yield _frame(s)
-                    # Stream ended ([DONE] or close): flush, then blank-bubble net.
-                    for s in buf.finish():
-                        yield _frame(s)
-                    # Safety net: LLM emitted ONLY [SKILL:...] tags and we dropped
-                    # them all → blank bubble; send a graceful fallback (2026-04-27).
-                    if buf.visible_chars == 0:
-                        yield _frame(
-                            "I tried to use a tool that didn't apply here. "
-                            "Could you rephrase, or just ask me to write it directly?"
-                        )
-                    yield "data: [DONE]\n\n"
-                except Exception as e:
-                    yield f"data: {json.dumps({'error': str(e)})}\n\n"
-            from starlette.responses import StreamingResponse as _SR
-            return _SR(_stream_gen(), media_type="text/event-stream",
-                       headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no"})
-
-        # Non-streaming fallback (A-12 PR-3E-chat-stream): canonical codec_llm.call.
-        # raise_on_error=True preserves the original raise-on-failure (was an
-        # r.json() KeyError) → outer except → 500. codec_llm strips <think>; the
-        # `### FINAL ANSWER:` marker is dashboard-specific so it stays.
-        import re
-        answer = codec_llm.call(messages, **_common, raise_on_error=True)
-        answer = re.sub(r'###\s*FINAL ANSWER:\s*', '', answer).strip()
-
-        # ── Post-LLM skill routing ──
-        # If the LLM outputs [SKILL:name:query], execute and inline the result.
-        skill_tag = re.search(r'\[SKILL:(\w+):([^\]]+)\]', answer)
-        if skill_tag:
-            s_name, s_query = skill_tag.group(1), skill_tag.group(2)
-            if s_name in CHAT_SKILL_ALLOWLIST and not _budget.consume("post_llm_skill_tag"):
-                # re-audit medium: the non-streaming path previously skipped the
-                # step budget that the streaming _resolve_skill_tag enforces, so
-                # stream:false could run skills past the per-turn cap. Mirror the
-                # stream path: budget exhausted → drop the tag.
-                log.info("[Chat] step_budget exhausted — dropping [SKILL:...] tag (non-stream)")
-                answer = answer.replace(skill_tag.group(0), "")
-            elif s_name in CHAT_SKILL_ALLOWLIST:
-                try:
-                    _, s_result = await asyncio.to_thread(_try_skill_by_name, s_name, s_query)
-                    if s_result:
-                        answer = answer.replace(skill_tag.group(0), f"**{s_result}**")
-                except Exception as e:
-                    # A-22 fix: was a silent `pass` — if skill resolution blows
-                    # up, the raw [SKILL:...] tag leaks into the user's chat with
-                    # no footprint. Surface it (log + audit); behavior unchanged
-                    # (tag stays, chat still returns).
-                    log.warning(
-                        f"Post-LLM skill tag resolution failed for {s_name!r}: {e}")
-                    try:
-                        log_event(
-                            "post_llm_skill_tag_failed", source="codec-dashboard",
-                            message=f"Skill tag resolution failed: {s_name}",
-                            level="warning", outcome="error",
-                            extra={"skill": s_name, "error": str(e)[:200]},
-                        )
-                    except Exception:
-                        pass
-
-        return {"response": answer, "model": model}
-    except Exception as e:
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
+# H1 / SR-59: def _chat_vision_response → moved to routes/chat.py
+# H1 / SR-59: def _build_chat_system_prompt → moved to routes/chat.py
+# H1 / SR-59:  → moved to routes/chat.py
 # (Voice/WebSocket endpoints moved to routes/websocket.py)
 # (Agent endpoints moved to routes/agents.py)
